Splitting in submodules in preparation for interface/export design
