+++ amazonka-s3-encryption/src/Network/AWS/S3/Encryption.hs
@@ -1,12 +1,9 @@
-{-# LANGUAGE DataKinds                  #-}
-{-# LANGUAGE FlexibleInstances          #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE LambdaCase                 #-}
-{-# LANGUAGE OverloadedStrings          #-}
-{-# LANGUAGE RecordWildCards            #-}
-{-# LANGUAGE ScopedTypeVariables        #-}
-{-# LANGUAGE TupleSections              #-}
-{-# LANGUAGE TypeFamilies               #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RankNTypes        #-}
+{-# LANGUAGE RecordWildCards   #-}
+{-# LANGUAGE TupleSections     #-}
+{-# LANGUAGE TypeFamilies      #-}
 
 -- |
 -- Module      : Network.AWS.S3.Encryption
@@ -18,36 +15,94 @@
 --
 module Network.AWS.S3.Encryption
     (
-    -- * Errors
-
-    -- * Operations
-
-    -- ** PutEncryptedObject
-    -- , module PutEncryptedObject
-
-    -- ** GetEncryptedObject
-    -- , module GetEncryptedObject
 
     ) where
 
--- import           Network.AWS.S3.Encryption.GetObject
+import           Control.Arrow
+import           Control.Monad.Catch
+import           Control.Monad.Reader
+import           Control.Monad.Trans.AWS
+import           Control.Monad.Trans.Resource
+import qualified Data.Aeson.Types                       as Aeson
+import           Data.Coerce
+import           Data.Conduit
+import           Data.Proxy
+import           Network.AWS.Prelude                    hiding (coerce)
+import           Network.AWS.Response
+import           Network.AWS.S3
+import qualified Network.AWS.S3                         as S3
+import           Network.AWS.S3.Encryption.Decrypt
+import           Network.AWS.S3.Encryption.Encrypt
+import           Network.AWS.S3.Encryption.Envelope
+import           Network.AWS.S3.Encryption.Instructions
 import           Network.AWS.S3.Encryption.Types
---import           Network.AWS.S3.Encryption.PutObject
+import           System.IO
 
--- - Note about or enforce the chunkSize to be a multiple of 128?
--- - How to get this module to appear in the generated cabal file?
--- - How to customise materials, kms vs aes etc.
--- - FUTURE: getObject should be transparent, or, if instruction file is supported then not?
+-- | Set the key material used to encrypt/decrypt a block of actions.
+material :: (MonadReader r m, HasKeyEnv r) => Key -> m a -> m a
+material k = local (envKey .~ k)
 
--- target :: AWSRequest a => Lens' (Encryd a) a
--- target = lens _target (\s a -> s { _tar = a })
+encrypt :: (AWSConstraint r m, HasKeyEnv r)
+        => PutObject
+        -> m PutObjectResponse
+encrypt x = do
+    (a, _) <- encrypted x
+    send (set location Metadata a)
 
--- setChunkSize :: MonadResource m => Int -> Conduit ByteString m ByteString
--- setChunkSize n = vectorBuilderC n mapM_CE =$= mapC fromByteVector
+encryptInstructions :: (AWSConstraint r m, HasKeyEnv r)
+                    => PutObject
+                    -> m PutObjectResponse
+encryptInstructions x = do
+    (a, b) <- encrypted x
+    _      <- send a
+    send b
 
--- FIXME: somehow updating an existing files instructions/metadata?
+-- | Note about parallelism/concurrency, and encryption of parts. If you don't
+-- encrypt any of the parts then the entire thing is unencrypted!
+initiate :: (AWSConstraint r m, HasKeyEnv r)
+         => CreateMultipartUpload
+         -> m ( CreateMultipartUploadResponse
+              , UploadPart -> Encrypted UploadPart
+              )
+initiate x = do
+    (a, _) <- encrypted x
+    rs     <- send (set location Metadata a)
+    return (rs, encryptPart a)
 
--- initiateMultiPartUpload
--- uploadPart
--- getobject
--- putobject
+initiateInstructions :: (AWSConstraint r m, HasKeyEnv r)
+                     => CreateMultipartUpload
+                     -> m ( CreateMultipartUploadResponse
+                          , UploadPart -> Encrypted UploadPart
+                          )
+initiateInstructions x = do
+    (a, b) <- encrypted x
+    rs     <- send a
+    _      <- send b
+    return (rs, encryptPart a)
+
+decrypt :: (AWSConstraint r m, HasKeyEnv r)
+        => GetObject
+        -> m GetObjectResponse
+decrypt x = do
+    let (a, _) = decrypted x
+    Decrypted f <- send a
+    f Nothing
+
+decryptInstructions :: (AWSConstraint r m, HasKeyEnv r)
+                    => GetObject
+                    -> m GetObjectResponse
+decryptInstructions x = do
+    let (a, b) = decrypted x
+    Instructions g <- send b
+    Decrypted    f <- send a
+    g >>= f . Just
+
+-- | Given a request to execute, such as AbortMultipartUpload or DeleteObject,
+-- remove the relevant adjacent instructions file, if it exists.
+cleanup :: (AWSConstraint r m, RemoveInstructions a)
+        => a
+        -> m (Rs a)
+cleanup x = do
+    rs <- send x
+    _  <- send (deleteInstructions x)
+    return rs
